finish: can delete / update dog with loading state add: error handling

+++ src/app/dogs/dogs.component.ts
@@ -3,14 +3,21 @@ import { DogService } from '../dog.service';
 import { Dog } from '../types';
 import { Subscription } from "rxjs";
 
+type UpdatedDog = {
+  idx: number;
+  newName: string
+}
+
 @Component({
   selector: 'app-dogs',
   templateUrl: './dogs.component.html',
-  styleUrls: ['./dogs.component.css']
+  styleUrls: ['./dogs.component.css'],
 })
 export class DogsComponent {
   dogs: Dog[] = []
+  uptadingDog: UpdatedDog = { idx: -1, newName: "" }
   isLoading = true;
+
   private subscription: Subscription = new Subscription();
 
   constructor(private dogService: DogService) {
@@ -36,13 +43,27 @@ export class DogsComponent {
     );
   }
 
-  updateDog(dog: Dog): void {
+  async updateDog(dog: Dog): Promise<void> {
+    this.isLoading = true;
+    if (this.uptadingDog.newName === "") {
+      alert("Name cannot be empty !")
+      return;
+    }
+    const newDog: Dog = { ...dog, name: this.uptadingDog.newName ?? "" }
+    await this.dogService.updateDog(newDog).toPromise()
+    this.dogs[this.uptadingDog.idx] = newDog;
+    this.uptadingDog = { idx: -1, newName: "" }
+    this.isLoading = false;
   }
 
   async deleteDog(dog: Dog): Promise<void> {
+    const isOK = confirm("Are you sure to want to delete this dog ?")
+    if (!isOK) return;
     this.isLoading = true;
     await this.dogService.deleteDog(dog.id).toPromise()
+    const idxOfDeletedDog = this.dogs.findIndex((currentDog) => currentDog.id === dog.id)
     this.dogs = this.dogs.filter((currentDog) => currentDog.id !== dog?.id)
+    if (idxOfDeletedDog === this.uptadingDog.idx) this.uptadingDog = { idx: -1, newName: "" }
     this.isLoading = false;
   }
 }
